+++ packages/tests/all/jawis-util ViewExceptionCallStack 1.ja.tsx
@@ -1,5 +1,5 @@
 import { TestProvision } from "^jarun";
-import { parseTrace } from "^parse-stack";
+import { parseTrace } from "^parse-captured-stack";
 import { getHtmlEnzyme } from "^misc/node";
 
 import { errorData1, errorData2, getViewExceptionCallStack } from "../_fixture";